+++ src/command/commands/image.ts
@@ -6,7 +6,9 @@ const command: Command = {
   help: '!image <prompt> - Creates an image based on the prompt.',
   platforms: [Platform.Twitch, Platform.Discord],
   run: async (ctx, { services }) => {
-    const prompt = ctx.body?.trim();
+    let prompt = ctx.body?.trim();
+    let sendToDiscord = true;
+
     const user = ctx.tags['display-name'];
     if (!prompt) {
       ctx.botSpeak(`You need to provide a prompt.`);
@@ -17,11 +19,19 @@ const command: Command = {
       `@${user} Please give me a few moments while I draw your image.`
     );
 
+    if (prompt.includes('nodiscord')) {
+      sendToDiscord = false;
+      prompt = prompt.replace(/nodiscord/gi, '').trim();
+    }
+
     let url = '';
     const firstWord = prompt.split(' ')[0].toLowerCase();
 
     if (firstWord === 'nate') {
-      url = await services.openaiService.editImage(`./public/images/edits/nate.png`, prompt.replace(/nate/i, 'this person'));
+      url = await services.openaiService.editImage(
+        `./public/images/edits/nate.png`,
+        prompt.replace(/nate/i, 'this person')
+      );
     } else {
       url = await services.openaiService.createImage(prompt);
     }
@@ -33,10 +43,14 @@ const command: Command = {
       url
     });
 
-    if (CONFIG.discord.enabled && CONFIG.discord.galleryChannelId) {
+    if (
+      sendToDiscord &&
+      CONFIG.discord.enabled &&
+      CONFIG.discord.galleryChannelId
+    ) {
       const res = await fetch(url);
       const buffer = Buffer.from(await res.arrayBuffer());
-      await services.discordService.postImageToGallery(
+      services.discordService.postImageToGallery(
         `@${user} on Twitch used !image ${prompt}`,
         buffer
       );